No longer mark AbstractFactory as internal
If a factory is used and it requires a key other than 'orm_default'
you must construct and invoke the factory yourself. Doing so will
invoke methods marked as internal, causing tooling to erroneously
mark using this factory as invalid. This patch removes the internal
annotation, allowing the AbstractFactory to be used to customize
the configuration key without triggering internal method usage.

Fixes #139